fix groups view
