+++ magpie/api/__init__.py
@@ -10,6 +10,4 @@ def includeme(config):
     config.include('magpie.api.home')
     config.include('magpie.api.login')
     config.include('magpie.api.management')
-
-    config.add_route('version', '/version')
     config.scan()
