+++ magpie/api/api_generic.py
@@ -4,17 +4,6 @@ from magpie.api.api_rest_schemas import *
 from magpie import __meta__, db
 
 
-@VersionAPI.get(tags=[APITag], api_security=SecurityEveryoneAPI, response_schemas=Version_GET_responses)
-@view_config(route_name='version', request_method='GET', permission=NO_PERMISSION_REQUIRED)
-def get_version(request):
-    """
-    Version information of the API.
-    """
-    return valid_http(httpSuccess=HTTPOk,
-                      content={u'version': __meta__.__version__, u'db_version': db.get_database_revision(request.db)},
-                      detail=Version_GET_OkResponseSchema.description, contentType='application/json')
-
-
 @notfound_view_config()
 def not_found(request):
     content = get_request_info(request, default_msg=NotFoundResponseSchema.description)
